requirements.txt
yfinance==0.2.33
matplotlib==3.8.1
pandas==2.1.0
numpy==1.26.0
python-docx==0.8.12
langchain-openai==0.0.217
